refactor(team): extract the inbox relay cursor helpers into a coalesce policy module
The OpenCode member inbox relay decides two different things while it walks a
member's unread rows: what a single delivery does with the runtime, and which
rows a delivery covers plus where the walk continues. The second question is
pure policy over the unread array, but it lived inline in a file that is already
the largest one in the provisioning folder, so every future rule about "which
rows ride along with this one" would have to grow it further.

Move findNextUnreadUserMessageIndex verbatim into a new sibling module that owns
inbox relay cursor and coalescing policy. Behaviour is unchanged: the relay
imports the helper instead of declaring it, and the existing relay tests still
exercise it through relayOpenCodeMemberInboxMessagesWithPorts.

The module is deliberately a sibling rather than a section of the relay because
the policy only needs RelayInboxMessage, while the relay needs the ledger, the
delivery ports and the file watcher. Keeping the direction one-way (relay to
policy, never back) is what lets the next commits put the coalescing rules here
without an import cycle.
